+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 
 #####1.0.7
 * [添加] Collection/List/Set/Map扩展方法
+* [修复] JetContext是链条状，SpringMvcFunctions查找Errors时没有查找parent
+* [修复] 针对jetbrick-template新添加的JetContextAware接口精简代码
 
 #####1.0.6
 * [添加] java.lang.Integer(int)扩展方法upto(), downto()